# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      resource:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
  models.Role:
    properties:
      company_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_by:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      failed_attempts:
        type: integer
      force_password_change:
        type: boolean
      id:
        type: string
      last_login_at:
        type: string
      last_login_ip:
        type: string
      locked_at:
        type: string
      locked_by:
        type: string
      mfa_enabled:
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  service.AssignRolesRequest:
    properties:
      role_ids:
        items:
          type: string
        type: array
    required:
    - role_ids
    type: object
  service.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      role_ids:
        items:
          type: string
        type: array
    required:
    - email
    - name
    type: object
  service.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  service.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 12
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  service.UpdateUserRequest:
    properties:
      name:
        type: string
      status:
        enum:
        - active
        - deactivated
        - locked
        - pending
        type: string
    type: object
...
  utils.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:5000
info:
  contact:
    email: support@peoplehub.com
    name: PeopleHub Team
  description: Enterprise HR Management & Payroll System API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PeopleHub API
  version: "1.1"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request a password reset token (public)
      parameters:
      - description: Email
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Forgot password
      tags:
      - Auth
...
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using a token (public)
      parameters:
      - description: Token + new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - Auth
...
  /users:
    get:
      description: List all users with pagination and filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search by email or name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user and return the generated password
      parameters:
      - description: User creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Soft-delete a user (status set to deleted)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      description: Get a single user with assigned roles
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user name or status
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
  /users/{id}/reset-password:
    post:
      description: Force reset a user's password and return the new one
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin reset password
      tags:
      - Users
  /users/{id}/roles:
    get:
      description: Get roles assigned to a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user roles
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Replace all roles for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.AssignRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign roles to user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"